[#86] chore: 파일 분리
